requirements.txt
alembic==1.7.4
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
flake8==4.0.1
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
SQLAlchemy==1.4.25
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
